requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pillow>=9.0.0
scikit-image>=0.19.0
einops>=0.4.0
